Move common C extension module into a different C file so that the actual definitions can be more easily included in other C extensions.
