+++ lib/python/setup.py
@@ -1,7 +1,8 @@
 from setuptools import setup, find_packages, Extension
 
 common_module = Extension("common",
-                          sources=["common_extension.c"],
+                          sources=["common_extension_module.c",
+                                   "common_extension.c"],
                           include_dirs=["../../", "../../thirdparty"],
                           extra_objects=["../../build/libcommon.a"],
                           extra_compile_args=["--std=c99", "-Werror"])
